backend/requirements.txt
flask-ngrok
flask
torch
sentencepiece
python-dotenv
Pillow
timm==0.5.4
transformers==4.25.1
donut-python
gradio
langchain
llama-index